Adding error check for number of arguments passed through WithExpression and moving table function tests into engine_tests.go.
