+++ sql/errors.go
@@ -85,6 +85,10 @@ var (
 	// node or expression is called with an invalid number of arguments.
 	ErrInvalidChildrenNumber = errors.NewKind("%T: invalid children number, got %d, expected %d")
 
+	// ErrInvalidExpressionNumber is returned when the WithExpression method of a node
+	// is called with an invalid number of arguments.
+	ErrInvalidExpressionNumber = errors.NewKind("%T: invalid expression number, got %d, expected %d")
+
 	// ErrInvalidChildType is returned when the WithChildren method of a
 	// node or expression is called with an invalid child type. This error is indicative of a bug.
 	ErrInvalidChildType = errors.NewKind("%T: invalid child type, got %T, expected %T")
